+++ src/zinc/hal/stm32f4/pin.rs
@@ -94,11 +94,11 @@ impl PinConf {
     let gpreg = self.get_reg();
 
     let bits: u32 = match self.function {
-      GPIOOut => 0b01 << offset as uint,
-      GPIOIn  => 0b00 << offset as uint,
+      GPIOOut => 0b01 << offset as usize,
+      GPIOIn  => 0b00 << offset as usize,
       _       => unsafe { abort() },  // FIXME(farcaller): not implemented
     };
-    let mask: u32 = !(0b11 << offset as uint);
+    let mask: u32 = !(0b11 << offset as usize);
     let val: u32 = gpreg.MODER();
 
     gpreg.set_MODER(val & mask | bits);
@@ -106,19 +106,19 @@ impl PinConf {
 
   /// Sets output GPIO value to high.
   pub fn set_high(&self) {
-    let bit: u32 = 1 << self.pin as uint;
+    let bit: u32 = 1 << self.pin as usize;
     self.get_reg().set_BSRR(bit);
   }
 
   /// Sets output GPIO value to low.
   pub fn set_low(&self) {
-    let bit: u32 = 1 << (self.pin as uint + 16);
+    let bit: u32 = 1 << (self.pin as usize + 16);
     self.get_reg().set_BSRR(bit);
   }
 
   /// Returns input GPIO level.
   pub fn level(&self) -> ::hal::pin::GpioLevel {
-    let bit: u32 = 1 << (self.pin as uint);
+    let bit: u32 = 1 << (self.pin as usize);
     let reg = self.get_reg();
 
     match reg.IDR() & bit {
